Use a custom `QueryPlanner` implementation so that we do not leak the `distribute_plan` function to the outside
