+++ src/distributed_planner/mod.rs
@@ -1,15 +1,14 @@
 mod batch_coalescing_below_network_boundaries;
 mod distribute_plan;
 mod distributed_config;
-mod distributed_physical_optimizer_rule;
 mod insert_broadcast;
 mod network_boundary;
 mod plan_annotator;
+mod session_state_builder_ext;
 mod task_estimator;
 
-pub use distribute_plan::distribute_plan;
 pub use distributed_config::DistributedConfig;
-pub use distributed_physical_optimizer_rule::DistributedPhysicalOptimizerRule;
 pub use network_boundary::{NetworkBoundary, NetworkBoundaryExt};
+pub use session_state_builder_ext::SessionStateBuilderExt;
 pub(crate) use task_estimator::set_distributed_task_estimator;
 pub use task_estimator::{TaskCountAnnotation, TaskEstimation, TaskEstimator};
