+++ src/distributed_planner/session_state_builder_ext.rs
@@ -1,0 +1,52 @@
+use crate::DistributedExec;
+use crate::distributed_planner::distribute_plan::distribute_plan;
+use async_trait::async_trait;
+use datafusion::common::Result;
+use datafusion::execution::context::QueryPlanner;
+use datafusion::execution::{SessionState, SessionStateBuilder};
+use datafusion::logical_expr::LogicalPlan;
+use datafusion::physical_plan::ExecutionPlan;
+use datafusion::physical_planner::{DefaultPhysicalPlanner, PhysicalPlanner};
+use std::sync::Arc;
+
+pub trait SessionStateBuilderExt {
+    fn with_distributed_planner(self) -> Self;
+}
+
+impl SessionStateBuilderExt for SessionStateBuilder {
+    fn with_distributed_planner(mut self) -> Self {
+        let prev = std::mem::take(self.query_planner());
+        self.with_query_planner(Arc::new(DistributedQueryPlanner { prev }))
+    }
+}
+#[derive(Debug)]
+struct DistributedQueryPlanner {
+    prev: Option<Arc<dyn QueryPlanner + Send + Sync>>,
+}
+
+#[async_trait]
+impl QueryPlanner for DistributedQueryPlanner {
+    async fn create_physical_plan(
+        &self,
+        logical_plan: &LogicalPlan,
+        session_state: &SessionState,
+    ) -> Result<Arc<dyn ExecutionPlan>> {
+        let s_plan = match &self.prev {
+            None => {
+                // Use the default physical planner.
+                let planner = DefaultPhysicalPlanner::default();
+                planner
+                    .create_physical_plan(logical_plan, session_state)
+                    .await?
+            }
+            Some(prev) => {
+                prev.create_physical_plan(logical_plan, session_state)
+                    .await?
+            }
+        };
+        match distribute_plan(Arc::clone(&s_plan), session_state.config_options()).await? {
+            Some(d_plan) => Ok(Arc::new(DistributedExec::new(d_plan))),
+            None => Ok(s_plan),
+        }
+    }
+}
