+++ tests/tpcds_plans_test.rs
@@ -4,8 +4,7 @@ mod tests {
     use datafusion_distributed::test_utils::localhost::start_localhost_context;
     use datafusion_distributed::test_utils::{benchmarks_common, tpcds};
     use datafusion_distributed::{
-        DefaultSessionBuilder, DistributedExec, DistributedExt, assert_snapshot,
-        display_plan_ascii, distribute_plan,
+        DefaultSessionBuilder, DistributedExec, DistributedExt, assert_snapshot, display_plan_ascii,
     };
     use std::env;
     use std::fs;
@@ -10607,7 +10606,6 @@ mod tests {
 
         let df = d_ctx.sql(&query_sql).await?;
         let plan = df.create_physical_plan().await?;
-        let plan = distribute_plan(plan, d_ctx.copied_config().options()).await?;
         if plan.as_any().is::<DistributedExec>() {
             Ok(display_plan_ascii(plan.as_ref(), false))
         } else {
